requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML - Groq
groq==0.9.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Cache & Queue
redis==5.0.1
celery==5.3.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
tenacity==8.2.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
